tests/wpt/web-platform-tests/css/css-text/parsing/word-wrap-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-wrap with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-word-wrap">
<meta name="assert" content="word-wrap supports the full grammar 'normal | break-word | anywhere'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("word-wrap", "normal");
test_valid_value("word-wrap", "break-word");
test_valid_value("word-wrap", "anywhere");
</script>
</body>
</html>
